requirements.txt
python-telegram-bot
requests
json
